refactor: organize CLI apps
