--- apps/bash.ts
@@ -1,2 +1,0 @@
-export { BASH_SOURCE } from '../core/fs/bin';
-
