+++ apps/cli/bash.ts
@@ -1,0 +1,2 @@
+export { BASH_SOURCE } from '../../core/fs/bin';
+
