+++ apps/cli/init.ts
@@ -1,0 +1,2 @@
+export { INIT_SOURCE } from '../../core/fs/bin';
+
