+++ apps/cli/kill.ts
@@ -1,0 +1,2 @@
+export { KILL_SOURCE } from '../../core/fs/bin';
+
