+++ apps/cli/login.ts
@@ -1,0 +1,2 @@
+export { LOGIN_SOURCE } from '../../core/fs/bin';
+
