+++ apps/cli/ls.ts
@@ -1,0 +1,1 @@
+export { LS_SOURCE } from '../../core/fs/bin';
