+++ apps/cli/mkdir.ts
@@ -1,0 +1,1 @@
+export { MKDIR_SOURCE } from '../../core/fs/bin';
