+++ apps/cli/mv.ts
@@ -1,0 +1,1 @@
+export { MV_SOURCE } from '../../core/fs/bin';
