+++ apps/cli/nano.ts
@@ -1,0 +1,1 @@
+export { NANO_SOURCE } from '../../core/fs/bin';
