+++ apps/cli/ping.ts
@@ -1,0 +1,2 @@
+export { PING_SOURCE } from '../../core/fs/bin';
+
