+++ apps/cli/ps.ts
@@ -1,0 +1,1 @@
+export { PS_SOURCE } from '../../core/fs/bin';
