+++ apps/cli/rm.ts
@@ -1,0 +1,1 @@
+export { RM_SOURCE } from '../../core/fs/bin';
