+++ apps/cli/snapshot.ts
@@ -1,0 +1,2 @@
+export { SNAPSHOT_SOURCE } from '../../core/fs/bin';
+
