+++ apps/cli/src/bash.ts
