+++ apps/cli/src/browser.ts
