+++ apps/cli/src/cat.ts
