+++ apps/cli/src/kill.ts
