+++ apps/cli/src/ls.ts
