+++ apps/cli/src/ping.ts
