+++ apps/cli/src/reboot.ts
