+++ apps/cli/src/rm.ts
