+++ apps/cli/src/snapshot.ts
