+++ apps/cli/ulimit.ts
@@ -1,0 +1,2 @@
+export { ULIMIT_SOURCE } from '../../core/fs/bin';
+
