--- apps/desktop.ts
@@ -1,1 +1,0 @@
-export { DESKTOP_SOURCE } from '../core/fs/bin';
