+++ apps/examples/desktop.ts
@@ -1,0 +1,1 @@
+export { DESKTOP_SOURCE } from '../../core/fs/bin';
