+++ apps/examples/webBrowser.ts
@@ -1,0 +1,1 @@
+export { BROWSER_SOURCE } from '../../core/fs/bin';
