+++ apps/index.ts
@@ -1,20 +1,20 @@
-export * from './browser';
-export * from './sshClient';
-export * from './nano';
-export * from './webBrowser';
-export * from './ping';
-export * from './desktop';
-export * from './ls';
-export * from './mkdir';
-export * from './rm';
-export * from './mv';
-export * from './ps';
-export * from './kill';
-export * from './init';
-export * from './login';
-export * from './bash';
-export * from './snapshot';
-export * from './ulimit';
+export * from './examples/browser';
+export * from './examples/sshClient';
+export * from './examples/desktop';
+export * from './examples/webBrowser';
+
+export * from './cli/bash';
+export * from './cli/init';
+export * from './cli/kill';
+export * from './cli/login';
+export * from './cli/ls';
+export * from './cli/mkdir';
+export * from './cli/rm';
+export * from './cli/mv';
+export * from './cli/nano';
+export * from './cli/ping';
+export * from './cli/ps';
+export * from './cli/snapshot';
+export * from './cli/ulimit';
 
 export { BUNDLED_APPS } from '../core/fs/bin';
-
