--- apps/init.ts
@@ -1,2 +1,0 @@
-export { INIT_SOURCE } from '../core/fs/bin';
-
