--- apps/kill.ts
@@ -1,2 +1,0 @@
-export { KILL_SOURCE } from '../core/fs/bin';
-
