--- apps/login.ts
@@ -1,2 +1,0 @@
-export { LOGIN_SOURCE } from '../core/fs/bin';
-
