--- apps/ls.ts
@@ -1,1 +1,0 @@
-export { LS_SOURCE } from '../core/fs/bin';
