--- apps/ping.ts
@@ -1,2 +1,0 @@
-export { PING_SOURCE } from '../core/fs/bin';
-
