--- apps/rm.ts
@@ -1,1 +1,0 @@
-export { RM_SOURCE } from '../core/fs/bin';
