--- apps/snapshot.ts
@@ -1,2 +1,0 @@
-export { SNAPSHOT_SOURCE } from '../core/fs/bin';
-
