--- apps/ulimit.ts
@@ -1,2 +1,0 @@
-export { ULIMIT_SOURCE } from '../core/fs/bin';
-
