--- apps/webBrowser.ts
@@ -1,1 +1,0 @@
-export { BROWSER_SOURCE } from '../core/fs/bin';
